+++ packages/blocks-ui/CHANGELOG.md
@@ -1,5 +1,20 @@
 # blocks-ui
 
+## 0.0.78
+
+### Patch Changes
+
+- 1411594: Update prettier parser import for v2
+
+## 0.0.77
+
+### Patch Changes
+
+- 776b7c3: Bump deps, add resolutions for context libs
+- Updated dependencies [776b7c3]
+  - property-controls@0.0.59
+  - @blocks/react@0.0.63
+
 ## 0.0.76
 
 ### Patch Changes